% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{libro_rojo_especies}
\alias{libro_rojo_especies}
\title{Especies de Fauna Silvestre del Libro Rojo (SERFOR, 2018)}
\format{
Un tibble con 528 filas y 7 variables:
\describe{
  \item{libro_rojo_id}{Número correlativo.}
  \item{libro_rojo_categoria}{Nombre de la categoría de amenaza.}
  \item{libro_rojo_codigo}{Código de categoría ordenado (CR, EN, VU, NT, DD).}
  \item{grupo_taxonomico}{Grupo taxonómico (Anfibios, Aves, Invertebrados, Mamíferos, Reptiles).}
  \item{scientific_name_raw}{Nombre científico tal como figura en la publicación.}
  \item{sinonimo}{Sinónimo o nombre adicional.}
  \item{scientific_name}{Nombre científico normalizado.}
}
}
\source{
Servicio Nacional Forestal y de Fauna Silvestre (SERFOR, 2018).
}
\usage{
libro_rojo_especies
}
\description{
Lista de 528 especies de fauna silvestre amenazada analizadas en el
Libro Rojo de la Fauna Silvestre Amenazada del Perú.
}
\keyword{datasets}
