% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_name}
\alias{normalize_name}
\title{Normalizar nombres científicos}
\usage{
normalize_name(x)
}
\arguments{
\item{x}{Vector de caracteres con nombres científicos.}
}
\value{
Vector de caracteres normalizado.
}
\description{
Limpia espacios redundantes, signos de puntuación periféricos y asegura
formato de mayúscula inicial para el género y minúsculas para el epíteto.
}
\keyword{internal}
