% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_accents}
\alias{remove_accents}
\title{Remover tildes y diacríticos}
\usage{
remove_accents(x)
}
\arguments{
\item{x}{Cadena de texto o vector de caracteres.}
}
\value{
Cadena sin tildes ni acentos.
}
\description{
Remover tildes y diacríticos
}
\keyword{internal}
